+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # Changelog
 
 - Modify maths functions to accept both strings and numbers — strings will be coerced into numbers before applying the correct operation.
+- Add `roundTo` function as alias to `round` to mimic JS behaviour.